Now drawing QR codes!
Alignment patters are automatically added, which is one step closer to
the goal :D zpng is used to print the QR code as a pgn file. It can
also be rendered in a text file.
For now, unfilled modules are shown in intermediate gray.

A little bit of refactoring and renaming in qr-generator. more tidying
is still needed.